No more DoFace

+++ FDS/Base/Vector.h
@@ -34,7 +34,7 @@ struct Vector
 
 	inline float length()
 	{
-		return sqrt(x * x + y * y + z * z);
+		return sqrtf(x * x + y * y + z * z);
 	}
 
 	inline Vector& normalize() {
